bindings: deconstruct constructor and destructor
So far we detected, stored, mounted all mounted v1 cgroups, and opened file
descriptors for each mounted v1 controller in a __attribute__((constructor)) and
performed corresponding cleanup operations in a __attribute__((destructor)).
This was a bad idea since it meant that even an innocent

        lxcfs -h

to look at the help output would cause all aforementiond operations to be
performed. Let's not do this. We deconstruct the constructor and destructor and
perform all mentioned operations above in a dedicated fuse init (lxcfs_init) and
fuse destroy (lxcfs_destroy) function.
