+++ lxcfs.c
@@ -28,9 +28,25 @@
 #include <sys/epoll.h>
 #include <sys/mount.h>
 #include <sys/socket.h>
+#include <sys/syscall.h>
 
 #include "bindings.h"
 #include "config.h" // for VERSION
+
+/* Define pivot_root() if missing from the C library */
+#ifndef HAVE_PIVOT_ROOT
+static int pivot_root(const char * new_root, const char * put_old)
+{
+#ifdef __NR_pivot_root
+return syscall(__NR_pivot_root, new_root, put_old);
+#else
+errno = ENOSYS;
+return -1;
+#endif
+}
+#else
+extern int pivot_root(const char * new_root, const char * put_old);
+#endif
 
 void *dlopen_handle;
 
@@ -686,6 +702,355 @@ int lxcfs_chmod(const char *path, mode_t mode)
 	return -ENOENT;
 }
 
+/*
+ * Functions needed to setup cgroups in the __constructor__.
+ */
+
+#define ALLOC_NUM 20
+static bool store_hierarchy(struct hierarchies **h, char *stridx, char *c)
+{
+	if ((*h)->nctrl % ALLOC_NUM == 0) {
+		size_t n = ((*h)->nctrl / ALLOC_NUM) + 1;
+		n *= ALLOC_NUM;
+		char **tmp = realloc((*h)->ctrl, n * sizeof(char *));
+		if (!tmp) {
+			fprintf(stderr, "Out of memory\n");
+			exit(1);
+		}
+		(*h)->ctrl = tmp;
+	}
+
+	do {
+		(*h)->ctrl[(*h)->nctrl] = strdup(c);
+	} while (!(*h)->ctrl[(*h)->nctrl]);
+	(*h)->nctrl++;
+	return true;
+}
+
+static void print_subsystems(const struct hierarchies *h)
+{
+	int i;
+
+	fprintf(stderr, "hierarchies:\n");
+	for (i = 0; i < h->nctrl; i++) {
+		if (h->ctrl[i])
+			fprintf(stderr, " %d: %s: fd: %d\n", i, h->ctrl[i], h->ctrlfd[i]);
+	}
+}
+
+static bool mkdir_p(const char *dir, mode_t mode)
+{
+	const char *tmp = dir;
+	const char *orig = dir;
+	char *makeme;
+
+	do {
+		dir = tmp + strspn(tmp, "/");
+		tmp = dir + strcspn(dir, "/");
+		makeme = strndup(orig, dir - orig);
+		if (!makeme)
+			return false;
+		if (mkdir(makeme, mode) && errno != EEXIST) {
+			fprintf(stderr, "failed to create directory '%s': %s",
+				makeme, strerror(errno));
+			free(makeme);
+			return false;
+		}
+		free(makeme);
+	} while(tmp != dir);
+
+	return true;
+}
+
+static bool umount_if_mounted(void)
+{
+	if (umount2(BASEDIR, MNT_DETACH) < 0 && errno != EINVAL) {
+		fprintf(stderr, "failed to unmount %s: %s.\n", BASEDIR, strerror(errno));
+		return false;
+	}
+	return true;
+}
+
+static int pivot_enter(void)
+{
+	int ret = -1, oldroot = -1, newroot = -1;
+
+	oldroot = open("/", O_DIRECTORY | O_RDONLY);
+	if (oldroot < 0) {
+		fprintf(stderr, "%s: Failed to open old root for fchdir.\n", __func__);
+		return ret;
+	}
+
+	newroot = open(ROOTDIR, O_DIRECTORY | O_RDONLY);
+	if (newroot < 0) {
+		fprintf(stderr, "%s: Failed to open new root for fchdir.\n", __func__);
+		goto err;
+	}
+
+	/* change into new root fs */
+	if (fchdir(newroot) < 0) {
+		fprintf(stderr, "%s: Failed to change directory to new rootfs: %s.\n", __func__, ROOTDIR);
+		goto err;
+	}
+
+	/* pivot_root into our new root fs */
+	if (pivot_root(".", ".") < 0) {
+		fprintf(stderr, "%s: pivot_root() syscall failed: %s.\n", __func__, strerror(errno));
+		goto err;
+	}
+
+	/*
+	 * At this point the old-root is mounted on top of our new-root.
+	 * To unmounted it we must not be chdir'd into it, so escape back
+	 * to the old-root.
+	 */
+	if (fchdir(oldroot) < 0) {
+		fprintf(stderr, "%s: Failed to enter old root.\n", __func__);
+		goto err;
+	}
+	if (umount2(".", MNT_DETACH) < 0) {
+		fprintf(stderr, "%s: Failed to detach old root.\n", __func__);
+		goto err;
+	}
+
+	if (fchdir(newroot) < 0) {
+		fprintf(stderr, "%s: Failed to re-enter new root.\n", __func__);
+		goto err;
+	}
+
+	ret = 0;
+
+err:
+	if (oldroot > 0)
+		close(oldroot);
+	if (newroot > 0)
+		close(newroot);
+	return ret;
+}
+
+/* Prepare our new clean root. */
+static int pivot_prepare(void)
+{
+	if (mkdir(ROOTDIR, 0700) < 0 && errno != EEXIST) {
+		fprintf(stderr, "%s: Failed to create directory for new root.\n", __func__);
+		return -1;
+	}
+
+	if (mount("/", ROOTDIR, NULL, MS_BIND, 0) < 0) {
+		fprintf(stderr, "%s: Failed to bind-mount / for new root: %s.\n", __func__, strerror(errno));
+		return -1;
+	}
+
+	if (mount(RUNTIME_PATH, ROOTDIR RUNTIME_PATH, NULL, MS_BIND, 0) < 0) {
+		fprintf(stderr, "%s: Failed to bind-mount /run into new root: %s.\n", __func__, strerror(errno));
+		return -1;
+	}
+
+	if (mount(BASEDIR, ROOTDIR BASEDIR, NULL, MS_REC | MS_MOVE, 0) < 0) {
+		printf("%s: failed to move " BASEDIR " into new root: %s.\n", __func__, strerror(errno));
+		return -1;
+	}
+
+	return 0;
+}
+
+static bool pivot_new_root(void)
+{
+	/* Prepare new root. */
+	if (pivot_prepare() < 0)
+		return false;
+
+	/* Pivot into new root. */
+	if (pivot_enter() < 0)
+		return false;
+
+	return true;
+}
+
+static bool setup_cgfs_dir(void)
+{
+	if (!mkdir_p(BASEDIR, 0700)) {
+		fprintf(stderr, "Failed to create lxcfs cgroup mountpoint.\n");
+		return false;
+	}
+
+	if (!umount_if_mounted()) {
+		fprintf(stderr, "Failed to clean up old lxcfs cgroup mountpoint.\n");
+		return false;
+	}
+
+	if (unshare(CLONE_NEWNS) < 0) {
+		fprintf(stderr, "%s: Failed to unshare mount namespace: %s.\n", __func__, strerror(errno));
+		return false;
+	}
+
+	if (mount(NULL, "/", NULL, MS_REC | MS_PRIVATE, 0) < 0) {
+		fprintf(stderr, "%s: Failed to remount / private: %s.\n", __func__, strerror(errno));
+		return false;
+	}
+
+	if (mount("tmpfs", BASEDIR, "tmpfs", 0, "size=100000,mode=700") < 0) {
+		fprintf(stderr, "Failed to mount tmpfs over lxcfs cgroup mountpoint.\n");
+		return false;
+	}
+
+	return true;
+}
+
+static bool do_mount_cgroups(struct hierarchies **h)
+{
+	char *target;
+	size_t clen, len;
+	int i, ret;
+
+	for (i = 0; i < (*h)->nctrl; i++) {
+		char *controller = (*h)->ctrl[i];
+		clen = strlen(controller);
+		len = strlen(BASEDIR) + clen + 2;
+		target = malloc(len);
+		if (!target)
+			return false;
+		ret = snprintf(target, len, "%s/%s", BASEDIR, controller);
+		if (ret < 0 || ret >= len) {
+			free(target);
+			return false;
+		}
+		if (mkdir(target, 0755) < 0 && errno != EEXIST) {
+			free(target);
+			return false;
+		}
+		if (mount(controller, target, "cgroup", 0, controller) < 0) {
+			fprintf(stderr, "Failed mounting cgroup %s\n", controller);
+			free(target);
+			return false;
+		}
+
+		(*h)->ctrlfd[i] = open(target, O_DIRECTORY);
+		if ((*h)->ctrlfd[i] < 0) {
+			free(target);
+			return false;
+		}
+		free(target);
+	}
+	return true;
+}
+
+static bool cgfs_setup_controllers(struct hierarchies **h)
+{
+	if (!setup_cgfs_dir())
+		return false;
+
+	if (!do_mount_cgroups(h)) {
+		fprintf(stderr, "Failed to set up private lxcfs cgroup mounts.\n");
+		return false;
+	}
+
+	if (!pivot_new_root())
+		return false;
+
+	return true;
+}
+
+static int preserve_ns(int pid)
+{
+	int ret;
+	size_t len = 5 /* /proc */ + 21 /* /int_as_str */ + 7 /* /ns/mnt */ + 1 /* \0 */;
+	char path[len];
+
+	ret = snprintf(path, len, "/proc/%d/ns/mnt", pid);
+	if (ret < 0 || (size_t)ret >= len)
+		return -1;
+
+	return open(path, O_RDONLY | O_CLOEXEC);
+}
+
+void *lxcfs_init(struct fuse_conn_info *conn)
+{
+	FILE *f;
+	char *line = NULL;
+	size_t len = 0;
+	int i, init_ns = -1;
+	struct hierarchies *hierarchies;
+
+	if ((f = fopen("/proc/self/cgroup", "r")) == NULL) {
+		fprintf(stderr, "Error opening /proc/self/cgroup: %s\n", strerror(errno));
+		return NULL;
+	}
+
+	hierarchies = malloc(sizeof(*hierarchies));
+	if (!hierarchies)
+		return NULL;
+
+	while (getline(&line, &len, f) != -1) {
+		char *p, *p2;
+
+		p = strchr(line, ':');
+		if (!p)
+			goto out;
+		*(p++) = '\0';
+
+		p2 = strrchr(p, ':');
+		if (!p2)
+			goto out;
+		*p2 = '\0';
+
+		/* With cgroupv2 /proc/self/cgroup can contain entries of the
+		 * form: 0::/ This will cause lxcfs to fail the cgroup mounts
+		 * because it parses out the empty string "" and later on passes
+		 * it to mount(). Let's skip such entries.
+		 */
+		if (!strcmp(p, ""))
+			continue;
+
+		if (!store_hierarchy(&hierarchies, line, p))
+			goto out;
+	}
+
+	/* Preserve initial namespace. */
+	init_ns = preserve_ns(getpid());
+	if (init_ns < 0)
+		goto out;
+
+	hierarchies->ctrlfd = malloc(sizeof(int *) * hierarchies->nctrl);
+	if (!hierarchies->ctrlfd)
+		goto out;
+
+	for (i = 0; i < hierarchies->nctrl; i++)
+		hierarchies->ctrlfd[i] = -1;
+
+	/* This function calls unshare(CLONE_NEWNS) our initial mount namespace
+	 * to privately mount lxcfs cgroups. */
+	if (!cgfs_setup_controllers(&hierarchies))
+		goto out;
+
+	if (setns(init_ns, 0) < 0)
+		goto out;
+
+	print_subsystems(hierarchies);
+
+out:
+	free(line);
+	fclose(f);
+	if (init_ns >= 0)
+		close(init_ns);
+	return hierarchies;
+}
+void lxcfs_destroy(void* private_data)
+{
+	int i;
+	struct hierarchies *h = private_data;
+
+	for (i = 0; i < h->nctrl; i++) {
+		if (h->ctrl[i])
+			free(h->ctrl[i]);
+		if (h->ctrlfd && h->ctrlfd[i] >= 0)
+			close(h->ctrlfd[i]);
+	}
+	free(h->ctrlfd);
+	free(h->ctrl);
+	free(h);
+}
+
 const struct fuse_operations lxcfs_ops = {
 	.getattr = lxcfs_getattr,
 	.readlink = NULL,
@@ -721,8 +1086,8 @@ const struct fuse_operations lxcfs_ops = {
 	.releasedir = lxcfs_releasedir,
 
 	.fsyncdir = NULL,
-	.init = NULL,
-	.destroy = NULL,
+	.init = lxcfs_init,
+	.destroy = lxcfs_destroy,
 	.access = lxcfs_access,
 	.create = NULL,
 	.ftruncate = NULL,
